feat: restructure the wheel behavior
Previously, shift + mouse wheel triggered uniform zooming and
ctrl/⌘ + mouse wheel triggered horizontal zooming.

ctrl/⌘ + mouse wheel is better suited for the normal zooming behavior
(i.e., uniform zooming rather than horizontal zooming) since the input
pattern is generated for pinch-zooming using a trackpad as documented in
the MDN:
https://developer.mozilla.org/en-US/docs/Web/API/MouseEvent/ctrlKey

On the other hand, the horizontal zooming feature is less relevant now
because it overlaps with the pinch-by-drag, which is more powerful as it
can zoom non-uniformly in any direction.

So let's drop the horizontal zooming to free the ctrl/⌘ + mouse wheel
input pattern and to save some code, and repurpose the freed input
pattern for uniform zooming.