fix: fix density computation for sampling
If dataView.getMaxY() == dataView.getMinY(), it doesn't mean that the
data height is zero, but it is actually one because all data are placed
at the same vertical level instead of being absent.

Zero width and height corrupts succeeding computations by causing e.g.,
zero-division-error, so fix the calculation and avoid such corruption.